Remove default properties from AgentConfig
Also: validate moderation service URI in Program.cs (example 2 and 3)
